에디터 버블 메뉴 내에 쓰이는 클래스네임 생성이 안되었던 문제 고치기 (#429)
unocss 토큰 스캐너가 모든 .svelte 파일을 스캔하면서 unocss 유틸리티 이름과 매치되는 문자열들을 골라내 최종 css로 생성하는 구조라서
.svelte 파일만을 대상으로 돌아서 그 외 파일에 기입한 클래스네임은 최종 css에 포함되지 않아 파일 확장자를 svelte 로 바꾸게 되었습니다.
